pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ktroj_torchlight"
version = "0.1"
authors = [{name = "Karol Trojanowski", email = "trojanowski.ifuj@gmail.com"}]
readme = "README.md"
license = {text = "MIT License"}
description = "Tool for profiling torch models."
dependencies = [
    "torch>=2,<3",
    "torchvision>=0.18,<1",
    "pillow>=10"
]
requires-python = ">= 3.10"

[tool.flake8]
max-line-length = 160

[tool.autopep8]
max_line_length = 160
